cleanup testing_utils types & consts into dedicated files
