--- unit-tests/src/testing_utils/setup/mod.rs
@@ -1,1 +1,0 @@
-pub mod base_suite;
